air-backend/requirements.txt
openai
openai-agents
motor 
reportlab 
python-dotenv 
fastapi 
uvicorn
qrcode